+++ example/src/index.js
@@ -1,3 +1,3 @@
-const imgA = require("../img/jorik-kleen-793723-unsplash.jpg?resize=900x&gamma");
+const imgA = require("../img/jorik-kleen-793723-unsplash.jpg?resize=900x&gamma&flip&flop");
 const imgB = require("../img/jorik-kleen-793723-unsplash.jpg?resize=600x&format=png");
 const imgC = require("../img/jorik-kleen-793723-unsplash.jpg?resize=1200x&format=jpeg&quality=10");